Public export of Nexus code:   - aa0b1904a282af190340e0eecdd6080a03514048 by Andrew Taylor <ataylor@redtoad.ca>
GitOrigin-RevId: aa0b1904a282af190340e0eecdd6080a03514048

+++ components/nexus-cleanup/src/main/java/org/sonatype/nexus/cleanup/internal/rest/CleanupPolicyResource.java
@@ -311,12 +311,12 @@ public class CleanupPolicyResource
     }
   }
 
-  private static Integer toSeconds(final Integer days) {
+  private static Long toSeconds(final Long days) {
     if (days == null) {
       return null;
     }
     else {
-      return (int) TimeUnit.DAYS.toSeconds(days);
+      return TimeUnit.DAYS.toSeconds(days);
     }
   }
 }
